test(phase2): cut §14.2 new-code duplication by lifting StatusList helpers
SonarCloud quality gate on #1129 fired at 6.6% new-code duplication
(threshold 3%). Hotspots were `readBit`, `decodeStatusListJwt`, and the
gunzip+base64url payload-decode chain — copy-pasted between
`phase14_vc_revocation.test.ts` (existing) and the new
`phase14_did_deactivate_vc_cascade.test.ts`.

Lift the three pieces into the shared `__helpers__/setup.ts`:
  - `readStatusListBit(bitstring, index)`
  - `decodeStatusListJwt(jwt)`
  - `decodeStatusListBitstring(jwt)` (gunzip + base64url combined)

Both StatusList-bit tests now import the helpers and only keep their
own test-specific Express round-trip wrapper (`fetchListBitstring`).

Also pre-empt the Gemini PR review note on `wireOverlapKeys`: add an
explicit comment in the multi-key test's `beforeEach` documenting that
`setupAcceptanceTest()` does `container.reset()` + re-registers
production deps, so the bootstrap-fallback test always sees the
production `IssuerDidKeyRepositoryStub` regardless of which test ran
before it.
